Moved the User file to the routes directory
