+++ knexfile.js
@@ -1,20 +1,20 @@
-require("dotenv").config();
+require('dotenv').config();
 
 module.exports = {
   development: {
-    client: "pg",
+    client: 'pg',
     connection: {
-      host: "localhost",
-      user: "postgres",
-      database: "grants",
-      user: "postgres",
-      password: "ert39883988"
+      host: 'localhost',
+      user: 'postgres',
+      database: 'grants',
+      user: 'postgres',
+      password: 'ert39883988'
     },
     migrations: {
-      directory: "./database/migrations"
+      directory: './database/migrations'
     },
     seeds: {
-      directory: "./database/seeds"
+      directory: './database/seeds'
     }
   },
 
@@ -22,7 +22,7 @@ module.exports = {
     client: process.env.DB_STAGING_CLIENT,
     connection: {
       host: process.env.DB_STAGING_HOST,
-      database: "d6c3e15jegg184",
+      database: process.env.DB_STAGING,
       user: process.env.DB_STAGING_USER,
       password: process.env.DB_STAGING_PASSWORD,
       ssl: {
@@ -30,10 +30,10 @@ module.exports = {
       }
     },
     migrations: {
-      directory: "./database/migrations"
+      directory: './database/migrations'
     },
     seeds: {
-      directory: "./database/seeds"
+      directory: './database/seeds'
     }
   },
 
@@ -41,20 +41,19 @@ module.exports = {
     client: process.env.DB_PRODUCTION_CLIENT,
     connection: {
       host: process.env.DB_PRODUCTION_HOST,
-      database: "dad7aiuhqookm7",
+      database: 'dad7aiuhqookm7',
       user: process.env.DB_PRODUCTION_USER,
       password: process.env.DB_PRODUCTION_PASSWORD,
       ssl: {
         rejectUnauthorized: false
       }
     },
-    
-migrations: {
+
+    migrations: {
       directory: './database/migrations'
     },
     seeds: {
       directory: './database/seeds'
     }
   }
-
 };
